requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
distro==1.9.0
dj-database-url==2.3.0
Django==5.1.7
django-bootstrap5==25.1
django-htmx==1.22.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
gitdb==4.0.11
GitPython==3.1.41
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==2.2.3
openai==1.66.2
packaging==24.2
pandas==2.2.3
psycopg==3.2.5
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
